requirements.txt
Django==2.1.2
Pillow==5.2.0
djangorestframework==3.8.2
whitenoise==3.3.1
gunicorn==19.9.0
django-heroku
dj-database-url
psycopg2